Remove LineChecker and RowChecker methods from Form1
The implementations of the `LineChecker` and `RowChecker` methods have been removed. This includes the logic for checking image equality in `line1` and `roow1`, as well as the associated background color changes. The closing braces for the `RowChecker` method and the `Form1` class have also been deleted, along with the start of the `Form2` class definition.